webapi/webapi-iap-xwalk-tests/iap/InAppPurchase.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>IAP Test: InAppPurchase</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://crosswalk-project.org/#wiki/A-Possible-Web-IDL-for-IAP">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var key = 'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAsEVjLWKJ5B7EKjnH5NVx4eh83xntg1WXFIQhLXgtmiCU0Ro27dN8Q7Q5HIif4ugkFcQWFhlDC/WsliDKtLZQkadmS1sUweOanIo5+ZbNaQ1ITWwtkNdnPb+tRJCjhngft1hFIfLGsKrBMxS8slaWaQ537wgex74Pv5mxuQLhsFyei5GffzDQDz4KN6sQHPqtB+GLDWqKKk2bdAiopQRFRUwYR8edAYcHiMHHZSfidOjaYD1/timhFIUnwFihtvNy4K7OStjoy6kEwCmX9/zXS3SxKkOfk2bNg/Edc8xgym8GkfObUliySSNxS9ukzp1ShDuSjdwfJVt5mvmXk+7jiQIDAQAB';
iap.init(key);

[
  //attributes of the InAppPurchase interface
  ["function", "queryProductDetails"],
  ["function", "buy"],
  ["function", "restore"]
].forEach(function(attr) {
  var type = attr[0];
  var name = attr[1];

  test(function () {
    assert_true(name in iap, name + " attribute in InAppPurchase");
    assert_equals(typeof iap[name], type, name + " attribute of type");
  }, "Check if InAppPurchase." + name + " exists and type of " + type);
});

test(function () {
  iap.queryProductDetails(["gas", "yearly", "premium"]).then(
  function(products) {
    assert_not_equals(products, null, "queryProductDetails of InAppPurchase");
  },
  function(error) {
    assert_unreached("queryProductDetails request failed");
  });
}, "Check if InAppPurchase.queryProductDetails is valid");

test(function () {
  iap.buy("gas").then(
  function(transaction) {
    assert_not_equals(transaction, null, "purchase of InAppPurchase");
  },
  function(error) {
    assert_unreached("purchase request failed");
  });
}, "Check if InAppPurchase.purchase is valid");

test(function () {
  iap.restore().then(
  function(transactions) {
    assert_not_equals(transactions.length, 0, "restore of InAppPurchase");
  },
  function(error) {
    assert_unreached("restore request failed");
  });
}, "Check if InAppPurchase.restore is valid");

</script>
